+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".presentation.login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,11 +27,11 @@
         </activity>
 
         <activity
-            android:name=".ui.signup.SignUpActivity"
+            android:name=".presentation.signup.SignUpActivity"
             android:exported="false" />
 
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false" />
     </application>
